chromium/third_party/blink/web_tests/virtual/out-of-blink-cspee/external/wpt/content-security-policy/embedded-enforcement/subsumption_algorithm-none-expected.txt

This is a testharness.js-based test.
PASS Empty required csp subsumes empty list of returned policies.
PASS Empty required csp subsumes any list of policies.
PASS Empty required csp subsumes a policy with `none`.
FAIL Required policy that allows `none` does not subsume empty list of policies. assert_equals: expected (undefined) undefined but got (boolean) true
FAIL Required csp with effective `none` does not subsume a host source expression. assert_equals: expected (undefined) undefined but got (boolean) true
FAIL Required csp with `none` does not subsume a host source expression. assert_equals: expected (undefined) undefined but got (boolean) true
FAIL Required csp with effective `none` does not subsume `none` of another directive. assert_equals: expected (undefined) undefined but got (boolean) true
FAIL Required csp with `none` does not subsume `none` of another directive. assert_equals: expected (undefined) undefined but got (boolean) true
FAIL Required csp with `none` does not subsume `none` of different directives. assert_equals: expected (undefined) undefined but got (boolean) true
PASS Required csp with `none` subsumes effective list of `none`.
PASS Required csp with `none` subsumes effective list of `none` despite other keywords.
PASS Source list with exprssions other than `none` make `none` ineffective.
PASS Returned csp with `none` is subsumed by any required csp.
PASS Returned csp with effective `none` is subsumed by any required csp.
PASS Both required and returned csp are `none`.
FAIL Both required and returned csp are `none` for only one directive. assert_unreached: No message should be sent from the frame. Reached unreachable code
PASS Both required and returned csp are empty.
PASS Both required and returned csp are effectively 'none'.
Harness: the test ran to completion.
